Refactoring Contentmanager
