+++ unit3dup/media_manager/ContentManager.py
@@ -1,11 +1,15 @@
 # -*- coding: utf-8 -*-
-import multiprocessing
+import json
+import os
+import re
 
 from unit3dup.contents import Contents
 from unit3dup.contents import Media
 from unit3dup.automode import Auto
-from unit3dup.files import Files
-from common.custom_console import custom_console
+
+from common.utility.utility import ManageTitles
+from common.trackers.trackers import ITTData
+from common.mediainfo import MediaFile
 
 class ContentManager:
     def __init__(self, path: str, tracker_name: str, mode: str, force_media_type=None):
@@ -21,59 +25,158 @@ class ContentManager:
         self.mode = mode
         self.force_media_type = force_media_type
 
-    def get_files(self) -> list['Contents']:
-        """Based on selected mode"""
-        if self.mode in ["man", "folder"]:
-            # Manual call to load files from specified path
-            auto = Auto(path=self.path, mode=self.mode, tracker_name=self.tracker_name,force_media_type=self.force_media_type)
-            # run cli flag
-            file_list = auto.upload()
-            return self.process(file_list=file_list)
-        else:
-            # Automatic call to load files based on detected content..
-            auto = Auto(path=self.path, tracker_name=self.tracker_name, force_media_type=self.force_media_type)
-            # run cli flag
-            file_list = auto.scan()
-            return self.process(file_list=file_list)
+        self.languages: list[str] | None = None
+        self.display_name: str | None = None
+        self.meta_info_list: list[dict] = []
+        self.game_nfo: str = ''
+        self.category: int | None = None
 
+        self.meta_info: str | None = None
+        self.size: int | None = None
+        self.name: str | None = None
+        self.folder: str | None = None
+        self.file_name: str | None = None
+        self.torrent_path: str | None = None
+        self.doc_description: str | None = None
 
-    def process(self, file_list: list['Media']) -> list['Contents']:
-        """ process media files and return 'contents' """
+        self.tracker_name: str = tracker_name
+        self.path: str = os.path.normpath(path)
+        self.tracker_data = ITTData.load_from_module()
 
+        self.auto = Auto(path=self.path, mode=self.mode, tracker_name=self.tracker_name,
+                         force_media_type=self.force_media_type)
+        self.media_list = self.auto.upload() if self.mode in ["man", "folder"] else self.auto.scan()
+
+    def process(self)-> list['Contents']:
         contents = []
-        try:
-            with multiprocessing.Pool(processes=4) as pool:
-                contents = pool.map(self.create_content_from_media, file_list)
-        except KeyboardInterrupt:
-            # Try to clean...
-            custom_console.bot_error_log("Interrupted by the user. Exiting...")
-            pool.terminate()
-            pool.join()
-            custom_console.bot_warning_log("Processes ended")
-        except Exception as e:
-            custom_console.bot_error_log(f"Error: {e}. Please report it")
-            pool.terminate()
-            pool.join()
-        finally:
-            return contents
+        for media in self.media_list:
+            self.path = media.torrent_path
+            self.category = media.media_type
 
+            content = self.get_data(media=media)
+            if content:
+                contents.append(content)
+        return contents
 
-    def create_content_from_media(self, media: 'Media') -> Contents:
-        """Creates a `Contents` object for each media item"""
-        # Create content using the file or folder specified by the media object
-        files = Files(
-            path=media.torrent_path,
-            tracker_name=self.tracker_name,
-            guess_title = media.guess_title,
-            media_type=media.media_type,
+    def get_data(self, media: Media) -> Contents | bool:
+        """
+        Process files or folders and create a `Contents` object if the metadata is valid
+        """
+
+        if os.path.isdir(self.path):
+            self.process_folder()
+        elif os.path.isfile(self.path):
+            self.process_file()
+
+        if not self.meta_info:
+            return False
+        torrent_pack = bool(re.search(r"S\d+(?!.*E\d+)", self.path))
+
+        return  Contents(
             game_title=media.game_title,
             game_crew=media.crew,
             game_tags=media.game_tags,
+            guess_title=media.guess_title,
             season=media.guess_season,
             episode=media.guess_episode,
             screen_size=media.screen_size,
+            file_name=self.file_name,
+            folder=self.folder,
+            name=self.name,
+            size=self.size,
+            metainfo=self.meta_info,
+            category=self.category,
+            tracker_name=self.tracker_name,
+            torrent_pack=torrent_pack,
+            torrent_path=self.torrent_path,
+            display_name=self.display_name,
+            doc_description=self.doc_description,
+            audio_languages=self.languages,
+            game_nfo=self.game_nfo,
         )
 
-        # Get content data and return if valid
-        content = files.get_data()
-        return content if content else None
+
+    def process_file(self) -> bool:
+        """Process individual files and gather metadata"""
+        self.file_name = os.path.basename(self.path)
+        self.folder = os.path.dirname(self.path)
+        self.display_name, _ = os.path.splitext(self.file_name)
+        self.display_name = ManageTitles.clean(self.display_name)
+        self.torrent_path = os.path.join(self.folder, self.file_name)
+
+        # Process media info for language and metadata
+        media_info = MediaFile(file_path=self.path)
+        self.languages = media_info.available_languages
+        self.name = self.file_name
+
+        # test to check if it is a doc
+        self.doc_description = self.file_name
+        self._handle_document_category()
+
+        # Build meta_info
+        self.size = os.path.getsize(self.path)
+        self.meta_info = json.dumps([{"length": self.size, "path": [self.file_name]}], indent=4)
+        return True
+
+    def process_folder(self)-> bool:
+        """Process folder and gather metadata for a torrent containing multiple files"""
+        files_list = self.list_files_by_category()
+        if not files_list:
+            return False
+
+        self.file_name = files_list[0]
+        self.folder = self.path
+        self.display_name = ManageTitles.clean(os.path.basename(self.path))
+        self._set_languages_from_title_or_media()
+
+        self.torrent_path = self.folder
+        self.name = os.path.basename(self.folder)
+        self.doc_description = "\n".join(files_list)
+
+        # test to check if it is a doc
+        self._handle_document_category()
+
+        # Build meta_info
+        self.size = 0
+        self.meta_info_list = []
+        for file in files_list:
+            size = os.path.getsize(os.path.join(self.folder, file))
+            self.meta_info_list.append({"length": size, "path": [file]})
+            self.size += size
+            if file.lower().endswith(".nfo"):
+                self.game_nfo = os.path.join(self.folder, file)
+
+        self.meta_info = json.dumps(self.meta_info_list, indent=4)
+        return True
+    def list_files_by_category(self) -> list[str]:
+        """List files based on the content category"""
+        if self.category == self.tracker_data.category.get('game'):
+            return self.list_game_files()
+        return self.list_video_files()
+
+    def list_video_files(self) -> list[str]:
+        """List video files in the folder"""
+        return [file for file in os.listdir(self.path) if ManageTitles.filter_ext(file)]
+
+    def list_game_files(self) -> list[str]:
+        """List all files in a game folder"""
+        return [file for file in os.listdir(self.path)]
+
+    def _handle_document_category(self)-> None:
+        """Verify if it is a document"""
+        media_docu_type = ManageTitles.media_docu_type(self.file_name)
+        if media_docu_type:
+            self.category = self.tracker_data.category.get(media_docu_type)
+
+    def _set_languages_from_title_or_media(self) -> None:
+        """Set language from the title or media info"""
+        if not self.languages:
+            filename_split = self.display_name.upper().split(" ")
+            for code in filename_split:
+                if converted_code := ManageTitles.convert_iso(code):
+                    self.languages = [converted_code]
+                    break
+
+        if not self.languages:
+            media_info = MediaFile(file_path=os.path.join(self.folder, self.file_name))
+            self.languages = media_info.available_languages
